did not want to spend more time trying to port the demo project. keeping around for now as a refrence. scene is reset to basic
