+++ src/caveport/setAttributeForClass.js
